# high-load-systems/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CDN Control Plane API",
    "description": "\nControl plane for a content delivery network: the service that decides **where**\ncontent lives and **which** edge node a client should talk to. It never serves the\ncontent bytes itself.\n\n* **Node management** — registration, lifecycle and continuous telemetry ingest.\n* **Routing** — ranked edge-node selection by proximity, load and cache state.\n* **Distribution** — asset metadata and the regional policy governing where it may be cached.\n* **Invalidation** — tracked purge campaigns with per-node acknowledgement.\n\nErrors follow RFC 9457 and are returned as `application/problem+json`.\nEvery response carries `X-Instance-ID`, `X-Request-ID` and `X-Response-Time-ms`.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/assets": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "List assets",
        "operationId": "list_assets_api_v1_assets_get",
        "parameters": [
          {
            "name": "path_prefix",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "examples": [
                "/static/"
              ],
              "title": "Path Prefix"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AssetStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AssetRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "assets"
        ],
        "summary": "Register an asset",
        "operationId": "create_asset_api_v1_assets_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetRead"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current state of the resource.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assets/{asset_id}": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Get one asset",
        "operationId": "get_asset_api_v1_assets__asset_id__get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetRead"
                }
              }
            }
          },
          "404": {
            "description": "Referenced resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "assets"
        ],
        "summary": "Archive an asset and purge it from the edge",
        "description": "Archive the asset and return the purge campaign it triggered.\n\n202 with the purge event rather than a bare 204: removal from the edge is\nasynchronous, and the caller needs a handle to follow it to completion.",
        "operationId": "archive_asset_api_v1_assets__asset_id__delete",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Asset Id"
            }
          },
          {
            "name": "requested_by",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Requested By"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurgeRead"
                }
              }
            }
          },
          "404": {
            "description": "Referenced resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assets/{asset_id}/versions": {
      "post": {
        "tags": [
          "assets"
        ],
        "summary": "Publish a new version of an asset",
        "description": "Publish new content under an existing path.\n\nOne call produces three effects, atomically: the version is bumped, every replica\nof the old bytes is marked stale, and a purge campaign is opened against the nodes\nholding them. That chain is the reason this is a POST to a sub-collection rather\nthan a PATCH on the asset — it creates a new version resource and a new purge.",
        "operationId": "publish_version_api_v1_assets__asset_id__versions_post",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Asset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetVersionPublish"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetVersionPublished"
                }
              }
            }
          },
          "404": {
            "description": "Referenced resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current state of the resource.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assets/{asset_id}/distribution": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "Get an asset's distribution policy",
        "operationId": "get_distribution_api_v1_assets__asset_id__distribution_get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistributionPolicyRead"
                }
              }
            }
          },
          "404": {
            "description": "Referenced resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "assets"
        ],
        "summary": "Replace an asset's distribution policy",
        "operationId": "replace_distribution_api_v1_assets__asset_id__distribution_put",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Asset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DistributionPolicy"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DistributionPolicyRead"
                }
              }
            }
          },
          "404": {
            "description": "Referenced resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation or violated a business rule.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssetCreate": {
        "properties": {
          "origin_path": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "pattern": "^/\\S*$",
            "title": "Origin Path",
            "examples": [
              "/static/app.bundle.js"
            ]
          },
          "content_hash": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "pattern": "^[0-9a-f]{64}$",
            "title": "Content Hash",
            "description": "Lowercase hex SHA-256 of the artifact bytes."
          },
          "size_bytes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Size Bytes"
          },
          "content_type": {
            "type": "string",
            "maxLength": 127,
            "minLength": 1,
            "title": "Content Type",
            "examples": [
              "application/javascript"
            ]
          },
          "cache_ttl_seconds": {
            "type": "integer",
            "maximum": 31536000.0,
            "minimum": 0.0,
            "title": "Cache Ttl Seconds",
            "default": 3600
          },
          "publish": {
            "type": "boolean",
            "title": "Publish",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "origin_path",
          "content_hash",
          "size_bytes",
          "content_type"
        ],
        "title": "AssetCreate"
      },
      "AssetRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "origin_path": {
            "type": "string",
            "title": "Origin Path"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "cache_ttl_seconds": {
            "type": "integer",
            "title": "Cache Ttl Seconds"
          },
          "status": {
            "$ref": "#/components/schemas/AssetStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "origin_path",
          "content_hash",
          "size_bytes",
          "content_type",
          "version",
          "cache_ttl_seconds",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "AssetRead"
      },
      "AssetStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published",
          "archived"
        ],
        "title": "AssetStatus"
      },
      "AssetVersionPublish": {
        "properties": {
          "content_hash": {
            "type": "string",
            "maxLength": 64,
            "minLength": 64,
            "pattern": "^[0-9a-f]{64}$",
            "title": "Content Hash",
            "description": "Lowercase hex SHA-256 of the artifact bytes."
          },
          "size_bytes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Size Bytes"
          },
          "cache_ttl_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 31536000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Ttl Seconds"
          },
          "requested_by": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Requested By",
            "examples": [
              "ci-pipeline"
            ]
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "content_hash",
          "size_bytes",
          "requested_by"
        ],
        "title": "AssetVersionPublish",
        "description": "Publish new bytes under an existing origin path."
      },
      "AssetVersionPublished": {
        "properties": {
          "asset": {
            "$ref": "#/components/schemas/AssetRead"
          },
          "previous_version": {
            "type": "integer",
            "title": "Previous Version"
          },
          "replicas_invalidated": {
            "type": "integer",
            "title": "Replicas Invalidated"
          },
          "purge_event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Purge Event Id"
          }
        },
        "type": "object",
        "required": [
          "asset",
          "previous_version",
          "replicas_invalidated",
          "purge_event_id"
        ],
        "title": "AssetVersionPublished",
        "description": "Result of a version bump: the new asset state plus the purge it triggered."
      },
      "DistributionPolicy": {
        "properties": {
          "rules": {
            "items": {
              "$ref": "#/components/schemas/DistributionRuleItem"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Rules"
          }
        },
        "type": "object",
        "required": [
          "rules"
        ],
        "title": "DistributionPolicy",
        "description": "Full replacement of an asset's distribution rule set.\n\nModelled as PUT-with-full-body rather than per-rule POST/DELETE so the operation\nis idempotent: a CI pipeline can declare the desired policy repeatedly without\naccumulating duplicates or needing to diff against current state."
      },
      "DistributionPolicyRead": {
        "properties": {
          "rules": {
            "items": {
              "$ref": "#/components/schemas/DistributionRuleItem"
            },
            "type": "array",
            "title": "Rules"
          },
          "unrestricted": {
            "type": "boolean",
            "title": "Unrestricted",
            "description": "True when no enabled rule exists, meaning the asset may be cached in every region."
          }
        },
        "type": "object",
        "required": [
          "rules",
          "unrestricted"
        ],
        "title": "DistributionPolicyRead"
      },
      "DistributionRuleItem": {
        "properties": {
          "region_code": {
            "type": "string",
            "maxLength": 32,
            "minLength": 2,
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$",
            "title": "Region Code",
            "examples": [
              "eu-central"
            ]
          },
          "priority": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Priority",
            "default": 100
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "region_code"
        ],
        "title": "DistributionRuleItem"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_AssetRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AssetRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[AssetRead]"
      },
...
      "PurgeRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "scope": {
            "$ref": "#/components/schemas/PurgeScope"
          },
          "asset_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id"
          },
          "region_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region Id"
          },
          "path_pattern": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path Pattern"
          },
          "requested_by": {
            "type": "string",
            "title": "Requested By"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "status": {
            "$ref": "#/components/schemas/PurgeStatus"
          },
          "target_node_count": {
            "type": "integer",
            "title": "Target Node Count"
          },
          "acknowledged_count": {
            "type": "integer",
            "title": "Acknowledged Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "scope",
          "asset_id",
          "region_id",
          "path_pattern",
          "requested_by",
          "reason",
          "status",
          "target_node_count",
          "acknowledged_count",
          "created_at",
          "completed_at"
        ],
        "title": "PurgeRead"
      },
...
      "PurgeScope": {
        "type": "string",
        "enum": [
          "asset",
          "region",
          "global"
        ],
        "title": "PurgeScope"
      },
      "PurgeStatus": {
        "type": "string",
        "enum": [
          "pending",
          "in_progress",
          "completed",
          "partially_failed"
        ],
        "title": "PurgeStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Problem": {
        "title": "Problem",
        "description": "RFC 9457 problem details.",
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ]
      }
    }
  }
}